% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dataAccessors}
\alias{dataAccessors}
\title{Deck.gl data accessors}
\usage{
dataAccessors(...)
}
\arguments{
\item{...}{named accessors to be passed to the relevant deck.gl JavaScript Method.}
}
\description{
In deck.gl every layer type has a specific set of data accessors, see e.g.
those for \href{https://deck.gl/docs/api-reference/layers/scatterplot-layer#data-accessors}{ScatterPlotLayer}.
This function sets all defaults for all available layer functions in this
package.
}
\details{
Please refer to the relevant \href{https://deck.gl/docs/api-reference/layers/}{deck.gl documentation}
for a more detailed description of the available layer functions.
See Details for a list of currently available accessors, their defaults and the
layer types they apply to.

If you want to map a certain accessor to a data specific value, you will need to
add it to the data and provide the column name to the respective data accessor.

Currently, the following accessors are automatically set to he following defaults:
\itemize{
\item getRadius = 1 (ScatterplotLayer)
\item getColor = c(0, 0, 0, 255) (ScatterplotLayer, PathLayer)
\item getFillColor = c(0, 0, 0, 130) (ScatterplotLayer, PolygonLayer)
\item getLineColor = c(0, 0, 0, 255) (ScatterplotLayer, PolygonLayer)
\item getLineWidth = 1 (ScatterplotLayer, PolygonLayer)
\item getElevation = 1000 (PolygonLayer)
\item getWidth = 1 (PathLayer)
}

NOTE:
\itemize{
\item accessors \code{getPosition}, \code{getPath}, \code{getPolygon} are handled internally
and should not be set!
\item all \code{get*Color} accessors will accept either a vector of rgb(a) integers (0-255)
or a hex color string (potentially also with alpha) - see examples.
}
}
\examples{
# default accessors
dataAccessors()

# modify selected accessors
dataAccessors(
  getFillColor = c(0, 0, 255, 130),
  getLineColor = "#ff00ffaa"
)

}
