% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{popupOptions}
\alias{popupOptions}
\alias{tooltipOptions}
\title{Options for popups and tooltips}
\usage{
popupOptions(...)

tooltipOptions(...)
}
\arguments{
\item{...}{named options to be passed to the popups and tooltips of the map.
See \href{https://maplibre.org/maplibre-gl-js/docs/API/type-aliases/PopupOptions/}{maplibregl PopupOptions}
for details and available options.}
}
\description{
Options for popups and tooltips
}
\details{
Both \code{popupOptions} and \code{tooltipOptions} are passed to the PopupOptions object
of the \href{https://maplibre.org/maplibre-gl-js/docs/API/classes/Popup/}{maplibregl Popup}
constructor. See \href{https://maplibre.org/maplibre-gl-js/docs/API/type-aliases/PopupOptions/}{maplibregl PopupOptions}
for details.

The \code{popupOptions} and \code{tooltipOptions} in this package only differ in their
respective defaults. These are:

For \code{popupOptions}
\itemize{
\item anchor = "bottom"
\item className = ""
\item closeButton = TRUE
\item closeOnClick = TRUE
\item closeOnMove = FALSE
\item focusAfterOpen = TRUE
\item maxWidth = "none"
\item offset = 0
\item subpixelPositioning = FALSE
}

For \code{tooltipOptions}
\itemize{
\item anchor = "top-left"
\item className = "geoarrow-deckgl-tooltip"
\item closeButton = FALSE
\item closeOnClick = FALSE
\item closeOnMove = FALSE
\item focusAfterOpen = TRUE
\item maxWidth = "none"
\item offset = 0
\item subpixelPositioning = FALSE
}
}
\section{Functions}{
\itemize{
\item \code{popupOptions()}: options for popups

\item \code{tooltipOptions()}: options for tooltips

}}
\examples{
# default
popupOptions()
tooltipOptions()

# modify selected options
tooltipOptions(anchor = "bottom-right", className = "my-css-class-name")

}
